firebase_auth_project/urls.py
from django.contrib import admin
from django.urls import path, include
from django.contrib.staticfiles.storage import staticfiles_storage
from django.views.generic import RedirectView
from rest_framework.documentation import include_docs_urls

urlpatterns = [
    path('admin/', admin.site.urls),
    path('favicon.ico', RedirectView.as_view(url=staticfiles_storage.url('img/favicon.ico'))),

    path('api/v1/', include('client_api.urls')),
    path(r'docs/', include_docs_urls(title='Firebase_project_API')),
]
